Switch receipt email to SparkPost
