+++ lib/code_corps/sparkpost/sparkpost.ex
@@ -5,4 +5,5 @@ defmodule CodeCorps.SparkPost do
   defdelegate update_templates, to: Tasks
 
   defdelegate send_forgot_password_email(user, token), to: Emails
+  defdelegate send_receipt_email(charge, invoice), to: Emails
 end
